chg: dev: Bond status support for interface and port
Bond status checks the link_state and hw_bond_config
state on each interface and updates the bond status
to UP, DOWN or BLOCKED for the interface and for
the LAG port according to the bond state of each of
the interfaces in it.
Add component test and MD file.

Change-Id: If64c173f7874ebd8e85191550960c68c3317a539
Signed-off-by: Randall Esquivel <randall.esquivel@hpe.com>