chg: dev: Bond status support for interface and port
Bond status checks the link_state and hw_bond_config
state on each interface and updates the bond status
to UP, DOWN or BLOCKED for the interface and for
the LAG port according to the bond state of each of
the interfaces in it.
Add component test and MD file.

TG-967

Change-Id: I565d9e6c21f5c954f021b010ba6cdfb8fbe0a2cc
Signed-off-by: Agustin Meneses <agustin-jose.meneses-fuentes@hpe.com>